Refactor: Making Tulza Happy
